using int8_t as idx_t index and size limit saves significant memory on tiny systems, guess that having 128 items on a collection might be ok, except for systems like sdcard... TODO make idx_t nigger and parameterize NavPos

+++ src/menu/base.h
@@ -7,7 +7,7 @@
 */
 
 
-using idx_t=int;
+using idx_t=int8_t;
 
 enum class Roles {Panel,Menu,Title,Body,Prompt,Index,Cursor,Name,Mode,Value,Unit};
 enum class Modes {Normal,Edit,Tune};
